requirements.txt
anyio==3.3.4
asgiref==3.4.1
bitnob==1.0.4
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
orjson==3.6.4
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
qrcode==7.3.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
